1- fix addProgrammer(pCu not saved in database ) 2- change Duration from minutes to hours 3- change Deviation to double 4- create getFailedTasksForProject function 5- create getProgress function 6- create countStatusForProject function 7- create getFailedTasksForProgrammer function 8- create getFailedTasksForProgrammerInProject function 9- create calculateProgrammerProductivity function 10- create getSkillGap function 11- create getMaxMinDeviation function 12- create countStatusForProgrammerInProject function 13- create countStatusForProgrammer function

+++ database/migrations/2019_02_05_225937_create_programmers_table.php
@@ -26,15 +26,15 @@ class CreateProgrammersTable extends Migration
             $table->string('location')->nullable();
             $table->string('age')->nullable();
             $table->Integer('PMid')->references('id')->on('User')->onDelete('cascade') ; // foreign key for the Project Manager
-            $table->integer('pStr')->nullable();
-            $table->integer('pJud')->nullable();
-            $table->integer('pCu')->nullable();
-            $table->integer('pTech')->nullable();
+            $table->double('pStr')->nullable();
+            $table->double('pJud')->nullable();
+            $table->double('pCu')->nullable();
+            $table->double('pTech')->nullable();
             $table->integer('numOfTasks')->default(0);
             $table->integer('pStrSum')->default(0);
             $table->integer('pJudSum')->default(0);
             $table->integer('pCuSum')->default(0);
-            $table->integer('pTechSum')->default(0);
+            $table->double('pTechSum')->default(0.0);
             $table->integer('failedTasks')->default(0);
             $table->rememberToken();
             $table->timestamps();
